feat: merge upstream (#74)
 *feat: merge upstream

---------

+++ taiko-image/build-client-image.sh
@@ -2,7 +2,6 @@
 
 set -e
 
-# source ./common.sh
 tmp_dir=/mnt/disks/data/tmp
 
 client_branch="main"
@@ -15,9 +14,16 @@ if [[ "${REPO}" == "taikoxyz/taiko-client" ]]; then
 fi
 
 build_client_image() {
-    client_dir="${tmp_dir}/taiko-client"
-    rm -fr "${client_dir}"
-    git clone https://github.com/taikoxyz/taiko-client.git "${client_dir}"
+    if [[ "${debug}" == "true" ]]; then
+        client_dir=$(
+            cd ../taiko-client
+            pwd
+        )
+    else
+        client_dir="${tmp_dir}/taiko-client"
+        rm -fr "${client_dir}"
+        git clone https://github.com/taikoxyz/taiko-client.git "${client_dir}"
+    fi
     cd "${client_dir}"
     git checkout "${client_branch}"
     docker build -t taiko-client:local .
